+++ src/Invoice.php
@@ -4,8 +4,12 @@ namespace Laravel\Cashier;
 
 use Carbon\Carbon;
 use Dompdf\Dompdf;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\View;
+use Laravel\Cashier\Exceptions\InvalidInvoice;
+use Stripe\Customer as StripeCustomer;
 use Stripe\Invoice as StripeInvoice;
+use Stripe\InvoiceLineItem as StripeInvoiceLineItem;
 use Symfony\Component\HttpFoundation\Response;
 
 class Invoice
@@ -25,14 +29,34 @@ class Invoice
     protected $invoice;
 
     /**
+     * The Stripe invoice line items.
+     *
+     * @var \Stripe\Collection|\Stripe\InvoiceLineItem[]
+     */
+    protected $items;
+
+    /**
+     * The taxes applied to the invoice.
+     *
+     * @var \Laravel\Cashier\Tax[]
+     */
+    protected $taxes;
+
+    /**
      * Create a new invoice instance.
      *
      * @param  \Illuminate\Database\Eloquent\Model  $owner
      * @param  \Stripe\Invoice  $invoice
      * @return void
+     *
+     * @throws \Laravel\Cashier\Exceptions\InvalidInvoice
      */
     public function __construct($owner, StripeInvoice $invoice)
     {
+        if ($owner->stripe_id !== $invoice->customer) {
+            throw InvalidInvoice::invalidOwner($invoice, $owner);
+        }
+
         $this->owner = $owner;
         $this->invoice = $invoice;
     }
@@ -117,9 +141,7 @@ class Invoice
      */
     public function hasDiscount()
     {
-        return $this->invoice->subtotal > 0 &&
-            $this->invoice->subtotal != $this->invoice->total &&
-            ! is_null($this->invoice->discount);
+        return $this->rawDiscount() > 0;
     }
 
     /**
@@ -129,7 +151,25 @@ class Invoice
      */
     public function discount()
     {
-        return $this->formatAmount($this->invoice->subtotal + $this->invoice->tax - $this->invoice->total);
+        return $this->formatAmount($this->rawDiscount());
+    }
+
+    /**
+     * Get the raw discount amount.
+     *
+     * @return int
+     */
+    public function rawDiscount()
+    {
+        if (! isset($this->invoice->discount)) {
+            return 0;
+        }
+
+        if ($this->discountIsPercentage()) {
+            return (int) round($this->invoice->subtotal * ($this->percentOff() / 100));
+        }
+
+        return $this->rawAmountOff();
     }
 
     /**
@@ -151,7 +191,7 @@ class Invoice
      */
     public function discountIsPercentage()
     {
-        return $this->coupon() && isset($this->invoice->discount->coupon->percent_off);
+        return isset($this->invoice->discount) && isset($this->invoice->discount->coupon->percent_off);
     }
 
     /**
@@ -175,15 +215,25 @@ class Invoice
      */
     public function amountOff()
     {
-        if (isset($this->invoice->discount->coupon->amount_off)) {
-            return $this->formatAmount($this->invoice->discount->coupon->amount_off);
-        }
-
-        return $this->formatAmount(0);
+        return $this->formatAmount($this->rawAmountOff());
     }
 
     /**
-     * Get the tax total amount.
+     * Get the raw discount amount for the invoice.
+     *
+     * @return int
+     */
+    public function rawAmountOff()
+    {
+        if (isset($this->invoice->discount->coupon->amount_off)) {
+            return $this->invoice->discount->coupon->amount_off;
+        }
+
+        return 0;
+    }
+
+    /**
+     * Get the total tax amount.
      *
      * @return string
      */
@@ -193,13 +243,78 @@ class Invoice
     }
 
     /**
+     * Determine if the invoice has tax applied.
+     *
+     * @return bool
+     */
+    public function hasTax()
+    {
+        $lineItems = $this->invoiceItems() + $this->subscriptions();
+
+        return collect($lineItems)->contains(function (InvoiceLineItem $item) {
+            return $item->hasTaxRates();
+        });
+    }
+
+    /**
+     * Get the taxes applied to the invoice.
+     *
+     * @return \Laravel\Cashier\Tax[]
+     */
+    public function taxes()
+    {
+        if (! is_null($this->taxes)) {
+            return $this->taxes;
+        }
+
+        $this->refreshWithExpandedTaxRates();
+
+        return $this->taxes = collect($this->invoice->total_tax_amounts)
+            ->sortByDesc('inclusive')
+            ->map(function (object $taxAmount) {
+                return new Tax($taxAmount->amount, $this->invoice->currency, $taxAmount->tax_rate);
+            })
+            ->all();
+    }
+
+    /**
+     * Determine if the customer is not exempted from taxes.
+     *
+     * @return bool
+     */
+    public function isNotTaxExempt()
+    {
+        return $this->invoice->customer_tax_exempt === StripeCustomer::TAX_EXEMPT_NONE;
+    }
+
+    /**
+     * Determine if the customer is exempted from taxes.
+     *
+     * @return bool
+     */
+    public function isTaxExempt()
+    {
+        return $this->invoice->customer_tax_exempt === StripeCustomer::TAX_EXEMPT_EXEMPT;
+    }
+
+    /**
+     * Determine if reverse charge applies to the customer.
+     *
+     * @return bool
+     */
+    public function reverseChargeApplies()
+    {
+        return $this->invoice->customer_tax_exempt === StripeCustomer::TAX_EXEMPT_REVERSE;
+    }
+
+    /**
      * Get all of the "invoice item" line items.
      *
      * @return array
      */
     public function invoiceItems()
     {
-        return $this->invoiceItemsByType('invoiceitem');
+        return $this->invoiceLineItemsByType('invoiceitem');
     }
 
     /**
@@ -209,7 +324,7 @@ class Invoice
      */
     public function subscriptions()
     {
-        return $this->invoiceItemsByType('subscription');
+        return $this->invoiceLineItemsByType('subscription');
     }
 
     /**
@@ -218,19 +333,46 @@ class Invoice
      * @param  string  $type
      * @return array
      */
-    public function invoiceItemsByType($type)
+    public function invoiceLineItemsByType($type)
     {
-        $lineItems = [];
+        if (is_null($this->items)) {
+            $this->refreshWithExpandedTaxRates();
 
-        if (isset($this->lines->data)) {
-            foreach ($this->lines->data as $line) {
-                if ($line->type == $type) {
-                    $lineItems[] = new InvoiceItem($this->owner, $line);
-                }
-            }
+            $this->items = new Collection($this->invoice->lines->autoPagingIterator());
         }
 
-        return $lineItems;
+        return $this->items->filter(function (StripeInvoiceLineItem $item) use ($type) {
+            return $item->type === $type;
+        })->map(function (StripeInvoiceLineItem $item) {
+            return new InvoiceLineItem($this, $item);
+        })->all();
+    }
+
+    /**
+     * Refresh the invoice with expanded TaxRate objects.
+     *
+     * @return void
+     */
+    protected function refreshWithExpandedTaxRates()
+    {
+        if ($this->invoice->id) {
+            $this->invoice = StripeInvoice::retrieve([
+                'id' => $this->invoice->id,
+                'expand' => [
+                    'lines.data.tax_amounts.tax_rate',
+                    'total_tax_amounts.tax_rate',
+                ],
+            ], $this->owner->stripeOptions());
+        } else {
+            // If no invoice ID is present then assume this is the customer's upcoming invoice...
+            $this->invoice = StripeInvoice::upcoming([
+                'customer' => $this->owner->stripe_id,
+                'expand' => [
+                    'lines.data.tax_amounts.tax_rate',
+                    'total_tax_amounts.tax_rate',
+                ],
+            ], $this->owner->stripeOptions());
+        }
     }
 
     /**