+++ src/Notifications/ConfirmPayment.php
@@ -2,11 +2,11 @@
 
 namespace Laravel\Cashier\Notifications;
 
-use Laravel\Cashier\Payment;
 use Illuminate\Bus\Queueable;
-use Illuminate\Notifications\Notification;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Messages\MailMessage;
+use Illuminate\Notifications\Notification;
+use Laravel\Cashier\Payment;
 
 class ConfirmPayment extends Notification implements ShouldQueue
 {
@@ -60,10 +60,9 @@ class ConfirmPayment extends Notification implements ShouldQueue
         $url = route('cashier.payment', ['id' => $this->paymentId]);
 
         return (new MailMessage)
+            ->subject(__('Confirm Payment'))
             ->greeting(__('Confirm your :amount payment', ['amount' => $this->amount]))
             ->line(__('Extra confirmation is needed to process your payment. Please continue to the payment page by clicking on the button below.'))
-            ->action(__('Confirm Payment'), $url)
-            ->line(__('Thanks').',')
-            ->line(config('app.name'));
+            ->action(__('Confirm Payment'), $url);
     }
 }