+++ tests/Fixtures/User.php
@@ -2,11 +2,23 @@
 
 namespace Laravel\Cashier\Tests\Fixtures;
 
-use Laravel\Cashier\Billable;
-use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Model;
+use Illuminate\Notifications\Notifiable;
+use Laravel\Cashier\Billable;
 
 class User extends Model
 {
     use Billable, Notifiable;
+
+    public $taxRates = [];
+
+    /**
+     * Get the tax rates to apply to the subscription.
+     *
+     * @return array
+     */
+    public function taxRates()
+    {
+        return $this->taxRates;
+    }
 }