+++ tests/Unit/InvoiceTest.php
@@ -2,14 +2,16 @@
 
 namespace Laravel\Cashier\Tests\Unit;
 
-use Mockery as m;
 use Carbon\Carbon;
-use Stripe\Discount;
 use Carbon\CarbonTimeZone;
 use Laravel\Cashier\Invoice;
-use Laravel\Cashier\Tests\TestCase;
-use Stripe\Invoice as StripeInvoice;
 use Laravel\Cashier\Tests\Fixtures\User;
+use Laravel\Cashier\Tests\TestCase;
+use Mockery as m;
+use Stripe\Coupon;
+use Stripe\Customer as StripeCustomer;
+use Stripe\Discount;
+use Stripe\Invoice as StripeInvoice;
 
 class InvoiceTest extends TestCase
 {
@@ -23,8 +25,13 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_the_invoice_date()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->created = 1560541724;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $date = $invoice->date();
 
@@ -35,8 +42,13 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_the_invoice_date_with_a_timezone()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->created = 1560541724;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $date = $invoice->date('CET');
 
@@ -47,9 +59,14 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_total()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->total = 1000;
         $stripeInvoice->currency = 'USD';
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $total = $invoice->total();
 
@@ -59,9 +76,14 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_raw_total()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->total = 1000;
         $stripeInvoice->currency = 'USD';
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $total = $invoice->rawTotal();
 
@@ -71,10 +93,15 @@ class InvoiceTest extends TestCase
     public function test_it_returns_a_lower_total_when_there_was_a_starting_balance()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->total = 1000;
         $stripeInvoice->currency = 'USD';
         $stripeInvoice->starting_balance = -450;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $total = $invoice->total();
 
@@ -84,9 +111,14 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_subtotal()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->subtotal = 500;
         $stripeInvoice->currency = 'USD';
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $subtotal = $invoice->subtotal();
 
@@ -96,8 +128,13 @@ class InvoiceTest extends TestCase
     public function test_it_can_determine_when_the_customer_has_a_starting_balance()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->starting_balance = -450;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $this->assertTrue($invoice->hasStartingBalance());
     }
@@ -105,8 +142,13 @@ class InvoiceTest extends TestCase
     public function test_it_can_determine_when_the_customer_does_not_have_a_starting_balance()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->starting_balance = 0;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $this->assertFalse($invoice->hasStartingBalance());
     }
@@ -114,9 +156,14 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_starting_balance()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->starting_balance = -450;
         $stripeInvoice->currency = 'USD';
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $startingBalance = $invoice->startingBalance();
 
@@ -126,8 +173,13 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_raw_starting_balance()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->starting_balance = -450;
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $startingBalance = $invoice->rawStartingBalance();
 
@@ -136,11 +188,22 @@ class InvoiceTest extends TestCase
 
     public function test_it_can_determine_if_it_has_a_discount_applied()
     {
+        $coupon = new Coupon();
+        $coupon->amount_off = 50;
+
+        $discount = new Discount();
+        $discount->coupon = $coupon;
+
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->subtotal = 450;
         $stripeInvoice->total = 500;
-        $stripeInvoice->discount = new Discount();
-        $invoice = new Invoice(new User(), $stripeInvoice);
+        $stripeInvoice->discount = $discount;
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $this->assertTrue($invoice->hasDiscount());
     }
@@ -148,12 +211,45 @@ class InvoiceTest extends TestCase
     public function test_it_can_return_its_tax()
     {
         $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
         $stripeInvoice->tax = 50;
         $stripeInvoice->currency = 'USD';
-        $invoice = new Invoice(new User(), $stripeInvoice);
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
 
         $tax = $invoice->tax();
 
         $this->assertEquals('$0.50', $tax);
+    }
+
+    public function test_it_can_determine_if_the_customer_was_exempt_from_taxes()
+    {
+        $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
+        $stripeInvoice->customer_tax_exempt = StripeCustomer::TAX_EXEMPT_EXEMPT;
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
+
+        $this->assertTrue($invoice->isTaxExempt());
+    }
+
+    public function test_it_can_determine_if_reverse_charge_applies()
+    {
+        $stripeInvoice = new StripeInvoice();
+        $stripeInvoice->customer = 'foo';
+        $stripeInvoice->customer_tax_exempt = StripeCustomer::TAX_EXEMPT_REVERSE;
+
+        $user = new User();
+        $user->stripe_id = 'foo';
+
+        $invoice = new Invoice($user, $stripeInvoice);
+
+        $this->assertTrue($invoice->reverseChargeApplies());
     }
 }